+++ src/cli_utils.zig
@@ -13,11 +13,11 @@ pub fn takeBool(value: Value) bool {
         } else if (std.mem.eql(u8, string_value, "false")) {
             return false;
         } else {
-            terminal.printError("value is not a boolean.", .{});
+            terminal.printColored(&.{.red}, "value is not a boolean.", .{});
             std.process.exit(1);
         }
     } else {
-        terminal.printError("value is not a boolean.", .{});
+        terminal.printColored(&.{.red}, "value is not a boolean.", .{});
         std.process.exit(1);
     }
 }
@@ -26,7 +26,7 @@ pub fn takeString(value: Value) []const u8 {
     if (value == .string) {
         return value.string;
     } else {
-        terminal.printError("value is not a string.", .{});
+        terminal.printColored(&.{.red}, "value is not a string.", .{});
         std.process.exit(1);
     }
 }
@@ -35,7 +35,7 @@ pub fn takeNumber(value: Value) []const u8 {
     if (value == .number) {
         return value.number;
     } else {
-        terminal.printError("value is not a number.", .{});
+        terminal.printColored(&.{.red}, "value is not a number.", .{});
         std.process.exit(1);
     }
 }